Q2 SG1 FA #2
In this scenario, I am described as a programmer of a food delivery app. My audience is delivery customers who want to know their delivery charges before placing an order. The context I find myself in is a local delivery service setting pricing for various distances. I must create a function in order to calculate and return the total delivery fee, and that function must return the correct fee and work with any distance or rate input.

The code only consists of one function, which calculates and returns the total delivery fee. There are two parameters, "km" and "rate", these are for multiplying to get the total deliver fee. If ever the product is not applicable, it will display an error. Now, in the main program, it shall ask for two inputs, the distance, and the rate. After both are given, the function is called. There are two arguments, "km_input" and "rate_input", which are the variables that contain the data from the user's inputs. Finally, the total delivery fee is given, printed through an f-string.