2Q FA 3
In this scenario, I am a junior software developer designing a simple math utility, with my audience being my teachers and classmates. I will be working on a "Fun with Numbers" mini app that asks the user for their age and then calculates and displays the total sum of all numbers from 1 up to that age.

The code shall ask the user for their age in the form of an integer, and stores it in a variable named as "age" in the code. Another separate variable, named "count" holds the value of 0, and will increase based on the variable "age" through a loop. A for loop is used here, and uses the range function, starting from 1, up until the variable "age" plus 1. In that loop, the variable "count" will have its value changed. This is done through the code seen below.

for i in range(1,age+1):
    count += i

The operator "+=" is used as a shortcut, as without this the code will now become "count = count + i" making the code shorter.

Once the loop is satisfied, it shall perform the code below:

print(f"The sum of all numbers from 1 to {age} is: {count}")

As seen here, the output uses an f-string. If I were to input 3, the output will be shown as "The sum of all numbers from 1 to 3 is: 6"

If all of these are followed correctly, then the program runs successfully.