FA4 Q2
(Due to the problematic flowchart, the code will be slightly different.)

In this code, a nested for loop is used. Firstly, it asks for the amount of students and subjects. For the code to actually work, two separate variables are assigned the value of zero, to take note of the sum of the score of one student, before resetting to 0 for the next student, and to take note of the sum of the averages to calculate the class average. The first for loop makes use of the value of the amount of students, this will repeat based on how many students there are. In that for loop, is another, which focuses on the subjects, this will repeat based on how many subjects there are, while asking for the score of each subject. The inner loop ends here, then calculating the average of that specific student using the specific variable mentioned earlier that gets the sum of the scores, then being used in the average variable, and then is reset back to 0 for the next student. The value for the average is put into the other variable dedicated to contain the sum of the averages. Once all the loops have been satisfied, the class average is calculated using the sum of the averages and the amount of students.