FA 5 Q2
In this scenario, I am a developer attempting to develop a simple travel app. My audience is a user planning a vacation itinerary.

The user inputs five travel destinations. Later, they want to update the second and fifth destinations.

My goal is to develop a program that stores travel destinations in an array, allows updates to specific elements, and displays the updated list.

The program first imports the module "array" to give access to functions. An array is defined with no elements. The program then asks the user to enter 5 travel destinations. Using a for loop, a destination is inputted into a variable, then appended to the array using arr.append(location), and repeats until the range is satisfied (which is (1,6)). The original travel itinerary is then printed using another for loop in the array. This allows the program to print each element one by one, with a counter to properly number the destinations for both for loops. Then, the program asks the user to change their 2nd and 5th destinations. It first asks for the second destination, storing the input into a separate variable, then using arr.pop(1) to remove the element in the 1st index since we are trying to replace the second destination. After that, arr.insert(1,new_location2) is used to insert the value of the variable to the 1st index. The same procedure is done in replacing the 5th destination, this time with the indexes different and another variable. This time, the program requires index 4 to be popped and replaced. Again, using arr.pop(4), the 4th index is removed since we are trying to replace the fifth destination. After that, arr.insert(4,new_location5) is used to insert the value into the 4th index of the array. Once those are executed, another for loop is used to print the new updated travel itinerary. The for loop is used to print each element one by one, with a counter to number the list. Once all are executed, the code concludes.